# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODC Pratique API",
        "description": "API documentation for ODC Pratique project",
        "contact": {
            "email": "votre@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/v1/dettes": {
            "get": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Get list of debts",
                "description": "Retrieve a paginated list of debts with their associated clients and articles.",
                "operationId": "221e325d8bfb6aa51a5e3e2d55903f6b",
                "responses": {
                    "200": {
                        "description": "Debts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS"
                                            ]
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Dette"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "last_page_url": {
                                                    "type": "string"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Create a new debt",
                "description": "Create a new debt record and associate it with the given client and articles.",
                "operationId": "2b63b8e510eddacc9fb4b66d096fc357",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dette"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Debt created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS"
                                            ]
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ERROR"
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dettes/{id}": {
            "get": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Get debt details",
                "description": "Retrieve the details of a specific debt by its ID, including client and associated articles and payments.",
                "operationId": "0e84c3d33fbb4855fc03cdc7c709a6d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the debt to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Debt retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS"
                                            ]
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Debt not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ERROR"
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Delete a debt",
                "description": "Delete a specific debt by its ID.",
                "operationId": "0abe1f9a68d383d4ed077b50044837f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the debt to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Debt deleted successfully"
                    },
                    "404": {
                        "description": "Debt not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ERROR"
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Update a debt",
                "description": "Update the details of a specific debt by its ID.",
                "operationId": "f000443bf0ee241c41389e1ff67a462b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the debt to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dette"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Debt updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS"
                                            ]
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Debt not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ERROR"
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ERROR"
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Dette": {
                "required": [
                    "montant_total",
                    "date_echeance",
                    "statut",
                    "client_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "montant_total": {
                        "type": "number",
                        "format": "float"
                    },
                    "date_echeance": {
                        "type": "string",
                        "format": "date"
                    },
                    "statut": {
                        "type": "string"
                    },
                    "client_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}